Added "aidl3" variant of RPC protocol. JB#56344
Due to the following change in the upstream,

https://android.googlesource.com/platform/frameworks/native/+/d70160f2989f18d4209ea464eb1635d4cc7ff827%5E%21/#F0

The system and vendor binder have a different header in Android 11.
To adapt this change to Android 11, we have to introduce a new
"aidl3" variant of service manager in libgbinder, which target to
Android 11 or higher if the rpc protocol remains for the upcoming
major Android release.